Move Changes and Scores to All components
