+++ src/components/Choices.jsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import { PaperIlust, RockIlust, ScissorIlust } from '../../../Helper/ilusts';
+import { PaperIlust, RockIlust, ScissorIlust } from '../Helper/ilusts';
 
 const Choices = ({ choice }) => {
 	const show = [
